% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatacao.R
\name{formatar_numero}
\alias{formatar_numero}
\title{Formatar número}
\usage{
formatar_numero(x, acc = 0.1)
}
\arguments{
\item{x}{É um valor numérico.}

\item{acc}{É um valor numérico que define a acurácia do
número resultante.}
}
\value{
É uma string representando o número formatado.
}
\description{
Essa função formata números.
}
\examples{
formatar_numero(0.1)
}
