% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatacao.R
\name{formatar_porcentagem}
\alias{formatar_porcentagem}
\title{Formatar porcentagens}
\usage{
formatar_porcentagem(x, acc = 0.1, scale = 100)
}
\arguments{
\item{x}{É um valor numérico.}

\item{acc}{É um valor numérico que define a acurácia do
número resultante.}

\item{scale}{É o multiplicador do número resultante.}
}
\value{
É uma string representando o número formatado.
}
\description{
Essa função formata números em porcentagens.
}
\examples{
# Exemplos
formatar_porcentagem(0.1)
formatar_porcentagem(10, scale = 1)
}
