% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meses}
\alias{meses}
\title{Meses do ano}
\format{
An object of class \code{character} of length 12.
}
\usage{
meses
}
\description{
Objeto com os meses do ano em português.
}
\keyword{datasets}
